# contracts/openapi.yaml
components:
  schemas:
    AttachmentLink:
      additionalProperties: false
      properties:
        attachment_id:
          format: uuid
          type: string
        expected_revision:
          format: int64
          type: integer
      required:
      - attachment_id
      - expected_revision
      type: object
...
    CreateCall:
      additionalProperties: false
      properties:
        attachments:
          items:
            format: uuid
            type: string
          type: array
        confidence:
          format: int32
          minimum: 0
          type:
          - integer
          - 'null'
        criteria:
          items:
            $ref: '#/components/schemas/Criteria'
          type: array
        instrument:
          type:
          - string
          - 'null'
        market:
          type:
          - string
          - 'null'
        original_claimed_at:
          format: date-time
          type:
          - string
          - 'null'
        original_text:
          type: string
        path:
          type: string
        playbook_id:
          format: uuid
          type:
          - string
          - 'null'
        related_call:
          format: uuid
          type:
          - string
          - 'null'
        source_entry:
          type: string
        stance:
          type: string
        tags:
          items:
            format: uuid
            type: string
          type: array
        timeframe:
          type:
          - string
          - 'null'
      required:
      - original_text
      type: object
...
    IndexRequest:
      additionalProperties: false
      properties:
        model_id:
          type: string
      required:
      - model_id
      type: object
...
    TextInput:
      properties:
        text:
          type: string
      required:
      - text
      type: object
...
  securitySchemes:
    bearerAuth:
      scheme: bearer
      type: http
info:
  description: Modular Rust backend. UTC timestamps, decimal strings, immutable evidence,
    bounded history, resumable review drafts. See docs/status.md for acceptance limits.
  title: Scorebook Backend
  version: 0.3.0
openapi: 3.1.0
paths:
  /v1/attachments:
    post:
      operationId: post__v1_attachments
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                captured_at:
                  format: date-time
                  type: string
                file:
                  format: binary
                  type: string
                kind:
                  enum:
                  - scene
                  - supplement
                  - reference
                  - query
                  type: string
              required:
              - file
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/attachments/{id}:
    get:
      operationId: get__v1_attachments__id_
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            image/jpeg:
              schema:
                format: binary
                type: string
            image/png:
              schema:
                format: binary
                type: string
            image/webp:
              schema:
                format: binary
                type: string
          description: Original image bytes
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/attachments/{id}/index:
    post:
      operationId: post__v1_attachments__id__index
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndexRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/calls:
    get:
      operationId: get__v1_calls
      parameters:
      - in: query
        name: before
        required: false
        schema:
          format: date-time
          type:
          - string
          - 'null'
      - in: query
        name: cursor
        required: false
        schema:
          type:
          - string
          - 'null'
      - in: query
        name: instrument
        required: false
        schema:
          type:
          - string
          - 'null'
      - in: query
        name: limit
        required: false
        schema:
          format: int64
          type:
          - integer
          - 'null'
      - in: query
        name: market
        required: false
        schema:
          type:
          - string
          - 'null'
      - in: query
        name: q
        required: false
        schema:
          type:
          - string
          - 'null'
      - in: query
        name: tag
        required: false
        schema:
          type:
          - string
          - 'null'
      - in: query
        name: timeframe
        required: false
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
    post:
      operationId: post__v1_calls
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCall'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/calls/preview:
    post:
      operationId: post__v1_calls_preview
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextInput'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/calls/{id}:
    get:
      operationId: get__v1_calls__id_
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/calls/{id}/attachments:
    post:
      operationId: post__v1_calls__id__attachments
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachmentLink'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/calls/{id}/corrections:
    post:
      operationId: post__v1_calls__id__corrections
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Correction'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/calls/{id}/history:
    get:
      operationId: get__v1_calls__id__history
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: cursor
        required: false
        schema:
          type:
          - string
          - 'null'
      - in: query
        name: kind
        required: false
        schema:
          type:
          - string
          - 'null'
      - in: query
        name: limit
        required: false
        schema:
          format: int64
          type:
          - integer
          - 'null'
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/calls/{id}/outcome-revisions:
    post:
      operationId: post__v1_calls__id__outcome-revisions
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevisionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/calls/{id}/replays:
    post:
      operationId: post__v1_calls__id__replays
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationInput'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/calls/{id}/review-draft:
    get:
      operationId: get__v1_calls__id__review-draft
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
    post:
      operationId: post__v1_calls__id__review-draft
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftInput'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/calls/{id}/review-draft/discard:
    post:
      operationId: post__v1_calls__id__review-draft_discard
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscardDraft'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/calls/{id}/review-draft/publish:
    post:
      operationId: post__v1_calls__id__review-draft_publish
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishDraft'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/calls/{id}/review-reminder:
    post:
      operationId: post__v1_calls__id__review-reminder
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnoozeInput'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/calls/{id}/revisions:
    post:
      operationId: post__v1_calls__id__revisions
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCall'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/calls/{id}/void:
    post:
      operationId: post__v1_calls__id__void
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Change'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/capabilities:
    get:
      operationId: get__v1_capabilities
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/criteria:
    get:
      operationId: get__v1_criteria
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
  /v1/deletions:
    post:
      operationId: post__v1_deletions
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteConfirm'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data: {}
                  meta:
                    type: object
                required:
                - data
                - meta
                type: object
          description: Success
        '401':
          description: Authentication required
        '404':
          description: Not found in current tenant
        '409':
          description: Idempotency or revision conflict
        '422':
          description: Invalid input
...
security:
- bearerAuth: []